+++ application/packages/cron/src/feed/fetch-articles.ts
@@ -1,0 +1,88 @@
+import type Logger from '@trend-diary/common/logger'
+import type { ArticleMedia } from '@trend-diary/domain/article/media'
+import { err, type Result } from 'neverthrow'
+import type { FetchEnv } from '../env'
+import { storeArticles } from './article-store'
+import { FEED_CONFIGS, type FeedConfig, type NormalizedItem, normalizedItemSchema } from './config'
+import { fetchRssFeed } from './rss-client'
+
+// 1件の不正itemでメディア全体の取込を止めないよう、不正itemは警告ログを残してスキップする。
+function selectValidItems(
+  media: ArticleMedia,
+  items: NormalizedItem[],
+  logger: Logger,
+): NormalizedItem[] {
+  const validItems: NormalizedItem[] = []
+  let skippedCount = 0
+
+  for (const item of items) {
+    const parsed = normalizedItemSchema.safeParse(item)
+    if (!parsed.success) {
+      skippedCount += 1
+      logger.warn({
+        msg: 'cron feed item skipped: validation failed',
+        media,
+        url: typeof item?.url === 'string' ? item.url : undefined,
+        issues: parsed.error.issues.map((issue) => ({
+          path: issue.path.join('.'),
+          message: issue.message,
+        })),
+      })
+      continue
+    }
+    validItems.push(parsed.data)
+  }
+
+  if (skippedCount > 0) {
+    logger.warn({
+      msg: 'cron feed items skipped',
+      media,
+      skippedCount,
+      totalCount: items.length,
+    })
+  }
+
+  return validItems
+}
+
+async function fetchAndStore<RawItem>(
+  media: ArticleMedia,
+  config: FeedConfig<RawItem>,
+  env: FetchEnv,
+  logger: Logger,
+): Promise<Result<number, Error>> {
+  const itemsResult = await fetchRssFeed<RawItem>(config.url)
+  if (itemsResult.isErr()) return err(itemsResult.error)
+
+  const validItems = selectValidItems(media, itemsResult.value.map(config.mapItem), logger)
+  return storeArticles(media, validItems, env)
+}
+
+export function fetchQiitaArticles(env: FetchEnv, logger: Logger): Promise<Result<number, Error>> {
+  return fetchAndStore('qiita', FEED_CONFIGS.qiita, env, logger)
+}
+
+export function fetchZennArticles(env: FetchEnv, logger: Logger): Promise<Result<number, Error>> {
+  return fetchAndStore('zenn', FEED_CONFIGS.zenn, env, logger)
+}
+
+export function fetchHatenaArticles(env: FetchEnv, logger: Logger): Promise<Result<number, Error>> {
+  return fetchAndStore('hatena', FEED_CONFIGS.hatena, env, logger)
+}
+
+const FETCHERS: Record<
+  ArticleMedia,
+  (env: FetchEnv, logger: Logger) => Promise<Result<number, Error>>
+> = {
+  qiita: fetchQiitaArticles,
+  zenn: fetchZennArticles,
+  hatena: fetchHatenaArticles,
+}
+
+export function runScheduledFetch(
+  media: ArticleMedia,
+  env: FetchEnv,
+  logger: Logger,
+): Promise<Result<number, Error>> {
+  return FETCHERS[media](env, logger)
+}
